+++ R/vdiffr.R
@@ -6,5 +6,8 @@
 
 wrapped_expect_doppelganger <- function(title, fig, path = NULL, ...) {
   testthat::skip_if_not_installed("vdiffr")
-  vdiffr::expect_doppelganger(title, fig, path = path, ...)
+  vdiffr::expect_doppelganger(
+    title, fig,
+    # path = path,
+    ...)
 }